北京理工大学/requirements.txt
altgraph==0.17.2
asgiref==3.6.0
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
barcode==1.0.2
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.8
colorama==0.4.4
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
Django==4.1.7
django-qrcode==0.3
entrypoints==0.4
executing==0.8.3
fonttools==4.39.2
future==0.18.2
idna==3.3
image==1.5.33
importlib-resources==5.12.0
ipykernel==6.13.0
ipython==8.3.0
itchat==1.3.10
itchat-uos==1.4.1
jedi==0.18.1
jieba==0.42.1
jsonpath==0.82
jupyter-client==7.3.1
jupyter-core==4.10.0
kiwisolver==1.4.4
Levenshtein==0.20.9
LinearRegression==0.0.1
lxml==4.6.4
matplotlib==3.7.1
matplotlib-inline==0.1.3
MouseInfo==0.1.3
nest-asyncio==1.5.5
numpy==1.21.4
opencv-python==4.5.4.60
packaging==21.3
pandas==1.4.2
parso==0.8.3
pefile==2021.9.3
pickleshare==0.7.5
Pillow==8.4.0
prettytable==3.4.1
prompt-toolkit==3.0.29
psutil==5.9.1
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
PyAutoGUI==0.9.50
pygal==3.0.0
PyGetWindow==0.0.9
Pygments==2.12.0
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
pypng==0.0.21
PyQRCode==1.2.1
PyRect==0.1.4
PyScreeze==0.1.28
python-barcode==0.14.0
python-dateutil==2.8.2
python-Levenshtein==0.20.9
pytweening==1.0.4
pytz==2022.1
pywin32==304
pywin32-ctypes==0.2.0
pyzbar==0.1.9
pyzmq==23.0.0
qrcode==7.4.2
rapidfuzz==2.13.7
requests==2.26.0
retry==0.9.2
retrying==1.3.4
rsa==4.8
six==1.16.0
sklearn==0.0.post1
soupsieve==2.3.1
sqlparse==0.4.3
stack-data==0.2.0
tornado==6.1
tqdm==4.62.3
traitlets==5.2.1.post0
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.7
useragent==0.1.1
wavfile==4.6.0
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
zipp==3.15.0
